Argument properties file encoding changes
Updated the LDAP SDK's support for reading argument values from a
properties file to be more in line with the Java properties file
format.  Although the LDAP SDK doesn't use the Java Properties
mechanism for reading argument properties files because it needs to
support multiple occurrences of the same properties), it now expects
the argument properties file to use the same type of encoding.  It
now expects that files are encoded in the ISO 8859-1 character set,
and it now supports Unicode escape sequences that consist of a
backslash followed by the letter u and four hexadecimal digits.