Add Nokia to references to Alcatel/Lucent
Updated the LDAP SDK to update all mentions of the Alcatel-Lucent
8661 Directory Server to also include Nokia ("Alcatel-Lucent" has
been replaced with "Nokia/Alcatel-Lucent") following the Nokia
acquisition of Alcatel-Lucent and the new product name.

Note that even though the product is now called the "Nokia 8661
Directory Server", references in the LDAP SDK still mention both
Nokia and Alcatel-Lucent because they apply to both new,
Nokia-branded versions of the product and older, Alcatel-Lucent
branded versions.