Fix incorrect result for unfollowable reference
Fixed an issue in which the LDAP SDK could override the result code
returned in a search result done message if the SDK was configured
to automatically follow referrals and it encountered a search result
reference that it could not automatically follow for some reason.
In this case, the LDAP SDK would make the unfollowable search result
reference available to the caller through the same mechanism that
would have been used if the SDK had not been configured to try to
automatically follow referrals (either by passing it off to a
SearchResultListener or collecting it in a list that would be
included in the SearchResult), but it could also override the result
code for the operation to be REFERRAL (10) instead of whatever
result code the server returned in the search result done message.
This would hide the actual result code that the server returned, and
would also cause the LDAP SDK to throw an exception (as it typically
does for non-success results).

With this change, the LDAP SDK no longer overrides the result code
for the search operation, so it will no longer throw an exception.
It will still make the unfollowable search reference available to
the caller through the SearchResultListener or via the SearchResult.