Change the order of modifications in Entry.diff
Updated the Entry.diff method to change the order in which
modifications are provided in the list that is returned.  When
operating in reversible mode and there is an attribute that exists
in both the source and target entries but with a different set of
values, the method would previously have included the add
modification with the newly added values before the delete
modification with the newly removed values.  This change flips that
order so that the delete modification comes before the add, which
should yield more reliable results, especially when changing the
case of a value in a case-insensitive attribute.