Add support for periodic pool maintenance
Updated the LDAPConnectionPoolHealthCheck class to add a new
performPoolMaintenance method that can be used to perform regular
maintenance on a connection pool as a whole (rather than tied to any
individual connection) at regular intervals specified by the pool's
health check interval.

Also added a PruneUnneededConnectionsLDAPConnectionPoolHealthCheck
class that uses the new performPoolMaintenance method to monitor the
size of the pool over time.  If the number of available connections
is consistently over a given minimum for at least a specified
duration, then the health check will shrink the pool to that
configured minimum.  This can be used to help reduce the size of a
pool during non-peak times when it may not be necessary to have the
maximum number of connections available.