Add more options for unparsable schema elements
Updated the Schema class to add additional options for handling
unparsable schema elements encountered while trying to decode an
entry.  Previously, any unparsable schema elements would be silently
ignored.  With this change, there is a new constructor that makes it
possible to get detailed information about all unparsable schema
elements.  There are also new static methods that can throw an
exception when trying to parse a schema entry, when trying to read
schema from an LDAP connection, or when trying to read schema from
files on the filesystem if any of the schema elements cannot be
parsed.