Add a StaticUtils.toArray method
Added a StaticUtils.toArray method that can be used to create an
array from a collection in a generic manner.  Although this is
usually not any more convenient than calling Collection.toArray, it
can be useful in certain contexts where  it isn't feasible to make
more than one call (for example, in a call to this or super in a
constructor) or where the type of collection element is not known
at compile time.