Make argument value validators serializable
Updated all existing ArgumentValueValidator instances to implement
the Serializable interface to avoid errors when trying to serialize
an argument configured with one of those validators (or when trying
to serialize an argument parser with any such arguments).  The
ArgumentValueValidator abstract class itself has not been altered to
maintain backward compatibility.