Work around an apparent JVM encryption issue
On some JVM versions, the PassphraseEncryptedInputStream and
PassphraseEncryptedOutputStream classes would occasionally
experience InvalidKeyExceptions thrown by
PassphraseEncryptedStreamHeader.  The underlying cause is that the
SecretKeyFactory.generateSecret method can return a different key
for the same input.  To work around this issue, the key is
generated multiple times and only used if the same key is
generated three times in a row.