Improve initial capacity values for maps and sets
Updated code that creates HashSet, LinkedHashSet, HashMap,
LinkedHashMap, and ConcurrentHashMap instances with an initial
capacity to use a better algorithm for computing that initial
capacity.  The new value now takes the load factor into account so
that it should be less likely to require re-hashing or re-balancing
the set or map, which should improve performance.

Also, added some StaticUtils convenience methods for creating
various types of sets with given contents.