Allow building from git with Java 7
Part of the LDAP SDK build process involves examining the workspace
in an attempt to obtain information about the version control
repository from which the workspace was checked out (including the
repository URL and revision identifier), which are inserted into the
dynamically generated com.unboundid.ldap.sdk.Version class.

Both subversion and git workspaces are supported, but git support is
implemented using the jgit library, and that library only works with
Java 8 or higher.  Previously, the attempt to build code checked out
from git would fail if built using Java SE 7.  This change allows
the build to succeed in that case, but it will use generic fallback
values for repository-related information.  It will print a warning
message about the inability to obtain repository information, but
the build will no longer fail.