Fix a bug in multivalued RDN normalization
Fixed a bug in generating the normalized representation of an RDN
with multiple values that have the same attribute type (for example,
"cn=foo+cn=bar").  In such cases, the normalized representation of
that RDN would have incorrectly had only one value with that
attribute type, and any other values with that same attribute type
would have been omitted.  Further, because the normalized string
representation of an RDN was used for other purposes (like
determining equality and comparator ordering), this may fix other
related issues as well.