Improve support for TLSv1.3
Updated the LDAP SDK to automatically enable support for TLSv1.3 in
JVMs that support it (which should be Java 11 or higher).  If
TLSv1.3 is supported in both the JVM in which the LDAP SDK is
running and in the server to which connections are being
established, then that protocol version will be preferred over
earlier versions.  If TLSv1.3 is not supported, then communication
will transparently fall back to the highest available TLS version
(at least out of TLSv1.2, TLSv1.1, or TLSv1; SSLv3 and earlier
protocol versions are not supported by default) that both the client
and server support.

Note that in testing TLSv1.3 support, I did encounter the OpenJDK
issue described at https://bugs.openjdk.java.net/browse/JDK-8211806.
This issue should be fixed in OpenJDK 11.0.2.  However, I decided
to enable support for TLSv1.3 anyway because it should only occur
under an unusual set of circumstances, and because the LDAP SDK's
TLS protocol support can be overridden programmatically or using
system properties if the problem does occur.  In my testing, the
problem only appeared under when all of the following are true:

* The client and the server both support TLSv1.3 and that
  protocol version is selected during negotiation.  The problem
  does not appear when the client and server have negotiated a
  lower protocol version (like TLSv1.2).

* The client and server have already established a TLS session
  from an earlier connection, and a new connection resumes that
  existing session instead of creating a new one.

* The client tried to get information about the certificates
  presented by the server during the negotiation process (e.g.,
  using SSLSession.getPeerCertificates) on a connection using the
  resumed session.

In such cases, the attempt to get information about the peer
certificates could fail with an SSLPeerUnverifiedException
indicating that the peer was unauthenticated.