Examine filesystem for fallback git repo details
The LDAP SDK build process tries to obtain information about the
repository (including the repository URL and revision ID) from which
the LDAP SDK was checked out so that it can be embedded in the
com.unboundid.ldap.sdk.Version source file, which is useful for
support purposes.  Both subversion and git repositories are
supported, but git support is more problematic than subversion
because the jgit library (which is our first choice for obtaining
information about the associated git repository) only works on Java
8 or later, while the LDAP SDK is supported on Java 7 and later.

We previously updated the build process so that if the attempt to
use jgit fails because of the Java version incompatibility, it will
fall back to trying to invoke the git command-line utility to obtain
the repository URL and revision, but even that wasn't good enough in
cases where the git command-line utility isn't available in the path
or a known location on the filesystem, or if an error occurred while
trying to invoke the command.  To make the build process even more
robust, this change adds another fallback option, which is to look
for the information in files in the .git directory tree.  If the
attempts to use both the jgit library and the git command-line tool
fail, the build process will now look in the .git/refs/heads/master
file for the revision ID and in the url property of the .git/config
file for the repository URL.