Improved tool support for password files
Added a new PasswordFileReader class that can be used to read
passwords from files.  The password reader ensures that the password
file contains exactly one line, and that the line is non-empty.  The
files may optionally be gzip-compressed and/or passphrase-encrypted,
and the password file reader will automatically detect and handle
them accordingly.  In the event that the file is encrypted, the key
may be obtained from a few different places (a Ping Identity
Directory Server's encryption settings database, pre-supplied, or
interactive user input).

All command-line tools now have access to a password file reader.
LDAP command-line tools use this if the bind password, keystore
password, or truststore password is to be read from a file.
The deliver one-time password and register YubiKey OTP device
tools use this if the user password is to be read from a file.  The
manage-certificates tool uses this if the keystore or private key
password is to be read from a file.