Improve DN and RDN validation
Updated the DN and RDN classes to add support for more strict
validation when creating or validating a DN or an RDN from a string.
The existing methods for creating and validating DNs and RDNs are
lenient with what they will allow in attribute names because some
servers are lenient and accept attribute names that are technically
invalid.  These methods still offer the same leniency for the sake
of backward compatibility, but new versions make it possible to
require attribute names to strictly comply with the specification.