Reject approx & extensible-match filters in DS
Updated the in-memory directory server to reject search requests
that include filters with approximate-match or extensible-match
components.  The LDAP SDK does not support client-side evaluation
for these filter types, but the in-memory directory server would
previously just not return any matching entries in response to a
search request using either of these filter types.  The new behavior
is more correct and makes it easier to troubleshoot unexpected
behavior for searches involving these filter types.