Make UNWILLING_TO_PERFORM no longer unusable
Updated the ResultCode.isConnectionUsable methods so that
UNWILLING_TO_PERFORM is no longer included in the set of result
codes that will cause the LDAP SDK to suspect that the connection
may no longer be usable.  Although it's possible that the connection
may have become invalid, there are plenty of reasons that an LDAP
server may return an UNWILLING_TO_PERFORM response for a
connection that remains completely usable.  Since isConnectionUsable
is often used to decide whether to keep the existing connection or
to throw it away and replace it with a new one, being too prone to
indicate that a connection is no longer usable can adversely impact
application performance and increase load on the directory server.