Immediately start TLS handshake on secure sockets
Updated the process for establishing a secure connection to
immediately start the TLS handshake on the socket, rather than
waiting for it to happen on the first attempt to communicate over
the connection.  This can help ensure that the connection is ready
to use more quickly, and can help avoid timing issues in certain
cases where the prompt trust manager is used in other interactive
applications that prompt for user input.