Use Ping DS's default name resolver if available
Updated the LDAPConnectionOptions class to choose the default name
resolver based on whether the LDAP SDK has access to the Ping
Identity Directory Server code base (which would indicate that it's
either running inside the server itself, or in a tool that ships
with the server).  If the LDAP SDK does have access to the server
code base, then it will use the server's default name resolver
implementation rather than an instance of the LDAP SDK's
DefaultNameResolver class.