Improve uniqueness control usability
Updated the support for the Ping Identity/UnboundID-proprietary
uniqueness request and response controls to help make it easier to
use them correctly.  The request control has been updated to include
an example in the class-level Javadoc documentation, and the
response control has been updated to add convenience methods to make
it easier to interpret the response.