Improve TLS connection failure messages
Improved the exception message that is generated if a failure occurs
while trying to create a TLS-based connection.  If the JVM supports
creating an unconnected SSLSocket and then connecting it after the
fact (which makes it possible to specify a connect timeout), and
that connection attempt failed (for example, because the client did
not trust the certificate presented by the server), the LDAP SDK
could think that the connection was still established.  Subsequent
attempts to use the connection would fail, but the failure message
would not accurately reflect the true cause of the problem.