Give ldapsearch and ldapmodify unlimited timeouts
Updated the ldapsearch and ldapmodify command-line tools to use an
unlimited response timeout, which will prevent the tool from giving
up on an operation if it takes the server a long time to return any
kind of response.  Previously, the tools used the LDAP SDK's default
timeout of 5 minutes for searches and 30 seconds for add, delete,
modify, and modify DN operations.