Make rate tools stoppable in executors
Updated the searchrate, modrate, authrate, and search-and-mod-rate
tools so that they can be programmatically stopped when the tool
instance is being run in an executor.  It was already possible to
programmatically stop the thread if it was running in a regular
thread, but this did not work as expected when the tool was run in
an executor.  The underlying reason is that the tool used the
Thread.join method to determine when the tool had completed, but
this doesn't work as intended when the tool is run in an executor
because the executor thread often doesn't exit, but is instead kept
around so that it could execute another Runnable.  The approach has
been redesigned so that it no longer depends on join, and it now
works as expected in either a background thread or in an executor.