Add a tool option to enable SSL/TLS debugging
Updated the LDAPCommandLineTool API to add an option to expose an
--enableSSLDebugging argument.  If this argument is available, and
if it is provided in the set of command-line arguments when the tool
is run, then the JVM's SSL/TLS debugging support will be enabled by
setting the "javax.net.debug" property to "all", and the JVM will
write a large amount of TLS-related debugging information to
standard error.  This can help troubleshoot problems with or provide
detailed information about any TLS communication that the tool
attempts.

The following tools have been updated to enable this argument:

* deliver-one-time-password
* deliver-password-reset-token
* dump-dns
* generate-source-from-schema
* generate-totp-shared-secret
* identify-references-to-missing-entries
* ldapcompare
* ldapmodify
* ldapsearch
* manage-account
* manage-certificates
* register-yubikey-otp-device
* subtree-accessibility
* validate-ldif