Update String creation from byte array
Updated the LDAP SDK to eliminate the last few references to code
in which a string was created from a byte array that did not go
through the StaticUtils.toUTF8String method.  Note that all of the
calls in the core LDAP SDK were already safe and could not have
resulted in the inadvertent creation of a string with an unexpected
encoding.  This update mainly includes unit tests, but does also
cover a couple of additional cases in the main LDAP SDK code where
a string was created from a byte array without going through the
StaticUtils.toUTF8String method.