Improve protection against blocked socket writes
Improved the LDAP SDK's protection against socket write attempts
that block for an indefinite length of time.  This is only likely to
occur when sending a large number of asynchronous requests over a
connection, and only in the case that the server stops reading
requests from the client, or if a networking problem prevents the
request from reaching the server and prevents the client from
receiving any information about that failure.

The LDAP SDK had previously attempted to protect against this
scenario through the use of the SO_TIMEOUT socket option on the
underlying connection, but it was discovered that this socket option
only applies to read operations, and that Java does not provide any
direct mechanism to enforce a timeout when writing to a blocking
socket.  The new approach uses a timer task to ensure that the
underlying connection is closed if a write attempt blocks for too
long, which will immediately cause the write attempt to fail.