Add LDAP URL support to searchrate
Updated the searchrate utility to allow specifying the base DN,
scope, filter, and requested attributes using LDAP URLs rather than
individual attribute values (note that any addresses and ports in
the URLs will be ignored; the --hostname and --port arguments will
still be used to identify which servers to use).  The LDAP URL can
be a fixed URL or it can be a value pattern (including the ability to
include variable content in the URLs, or to load URLs from a file).
Using LDAP URLs allows for more precise control over the combination
of base DN, scope, filter, and requested attributes on a per-request
basis.