Add an indent-filter tool
Added a new indent-filter command-line tool that can help make it
easier to visualize complex filters with a lot of components, and
especially a lot of nesting.  If possible, it can also try to
simplify the filter (for example, to remove unnecessary levels of
nesting, like an AND inside an AND).