Add a SubtreeDeleter utility class
Added a new SubtreeDeleter utility class that can make it easier
to delete a specified subtree, optionally including or excluding the
base entry for that subtree.  It provides a good client-side
alternative to the subtree delete request control, which isn't
supported by all servers, and can sometimes be problematic in
servers that do support it.

At its core, the subtree deleter works with a combination of search
and delete operations.  Although it can potentially operate using
only core LDAPv3 features, it can optionally take advantage of
several standard controls to make it more efficient and effective,
including:

* The simple paged results control
* The LDAP subentries control
* The manage DSA IT control

When run against a Ping Identity Directory Server, it can also
optionally take advantage of the following additional proprietary
features:

* The hard delete control
* The permit unindexed search control
* The return conflict entries control
* The soft-deleted entry access control
* The subtree accessibility extended operation

You can also provide additional controls for inclusion in search or
delete requests.  Further, it is possible to use a fixed-rate
barrier to impose a limit on the rate at which the deletes are
processed if that capability is desired.