Add client-side subtree delete to ldapmodify
Updated the ldapmodify command-line tool to add a
--clientSideSubtreeDelete argument that can be used to cause each
delete operation to be converted to a client-side subtree delete
operation, in which the tool will search for entries to delete and
then delete them individually.  This makes it easier to delete
entries with subordinates on servers that either do not support the
subtree delete request control, or in which the client may not
have permission to use that control.