Update the WriteWithTimeout implementation
Updated the WriteWithTimeout implementation to provide an optional Socket
argument.  If provided, the socket will be closed rather than the output
stream, which may help the closure to be more reliable in the event that the
attempt to close the output stream might trigger an attempt to write to what
we already believe to be a blocked stream.