Improve ByteStringBuffer append of CharSequence
Fixed a bug in the ByteStringBuffer.append(CharSequence,int,int)
method in which the final integer argument could be interpreted as
the number of characters to append rather than the end position at
which to stop appending, which could yield incorrect results when
the method was called with a nonzero start position.  Also, updated
the ByteStringBuffer.append methods that take CharSequence arguments
to eliminate the creation of an intermediate character array,
thereby improving performance and reducing garbage creation.