Add an ldapdelete command-line tool
Added an ldapdelete command-line tool that can be used to delete
entries from an LDAP directory server.  The DNs of the entries to
delete can be provided on the command line, read from a file, or
read from standard input.  Alternately, the server can search for
and delete all entries matching one or more filters.  It offers a
number of options, including support for client-side and server-side
subtree deletes, rate limiting, and a variety of standard and
proprietary controls.