Fix a name resolver initialization error
Fixed an error that could occur when initializing the default name
resolver in the LDAPConnectionOptions class.  By default, the LDAP
SDK's name resolver is used, but if the LDAP SDK detects that it's
running with a classpath that provides access to a Ping Identity
Directory Server classes, then it will instead try to use the
server's name resolver as the default.  However, if the server code
is available but an Error (that is, a non-Exception Throwable)
occurs while attempting to initialize it, the LDAPConnectionOptions
class initialization will fail.  This change avoids that scenario by
catching Throwable rather than Exception.