Limit default usage of a server name resolver
Updated the LDAPConnectionOptions code that selects the default name
resolver to be more restrictive about when it will actually try to
use the name resolver from a Ping Identity Directory Server (or
related product) as its default resolver.  Previously, it would use
the server's name resolver if the server code was available and the
resolver could be instantiated without throwing anything.  It will
now also check to make sure that the server instance root is defined
(either via a system property or an environment variable), that the
instantiated server name resolver returns non-null values from
getLocalHost and getByName (for the IP address obtained from the
local host), without throwing anything in the process.