Add custom interactive mode support to tools
Added a CommandLineTool.requestToolArgumentsInteractively method
that can be used to override the LDAP SDK's default interactive mode
processing for a tool.  If a tool supports interactive mode and
overrides this method, then the processing performed by that method
will be used to obtain the command-line arguments that should be
used to actually run the tool.