Improve the disallowed arg value error message
Improved the error message that would be returned by the argument
parser when the value provided for a string or Boolean argument was
not included in the set of allowed values for that argument.  It
would previously just tell you that the provided value was not
allowed.  It will now also provide a list of the allowed values.