Add generate password extended request/result
Added client-side support for the generate password extended request
and response, which may be used to request that the server generate
one or more passwords that may be suggested to an end user when
creating or updating an entry.  Support for this extended operation
will be included in an upcoming version of the Ping Identity
Directory Server.