Fix search-and-mod-rate search duration
Fixed an issue in the search-and-mod-rate tool in which the search
duration times reported included not only the time required to
process the search, but also the time required to process the
associated modify operations as well.  If rate limiting was used,
then the search duration would also include the wait before each
modify attempt.