Allow getting an env var with a default value
Added an additional StaticUtils.getEnvironmentVariable method that
allows you to provide a default value that should be used if the
requested environment variable is not set.

Also, updated the PasswordReader class to provide a way to obtain
the password from a file set through an environment variable.  This
is only intended for testing purposes and is not intended to be
used by most external clients.  Some no-longer-necessary legacy
code in the PasswordReader class has also been removed, but API
backward compatibility has been preserved.