Add working directory to the exec task
Updated support for the exec task to make it possible to specify the
path to the working directory to use when executing the command.
Previously, the server would always use the instance root directory
as the working directory, and that will still be the default if no
working directory is specified.