manage-certificates check usability improvement
Updated the manage-certificates check-certificate-usability command
to add an additional check to see whether the certificate at the
root of the chain is found in the JVM's set of trusted issuer
certificates.  If the CA certificate is not found in the JVM's
default trust store, a notice will be displayed, but the tool will
still complete with a success result.