Fix a debug issue if $COLUMNS is malformed
Fixed an issue that could arise if LDAP SDK debugging was enabled in
an environment in which the COLUMNS environment variable is not
available or has a malformed value.  The failure to parse the
COLUMNS environment variable resulted in an internal exception that
was being caught and sent to the debugging framework, but the debug
attempt needed access to the StaticUtils.UTC_TIME_ZONE constant,
which had not yet been assigned because that assignment happens after
the attempt to parse COLUMNS.  To address this problem, the attempt
to parse the COLUMNS environment variable now comes after the
UTC_TIME_ZONE constant is set.