Fix an X.509 certificate extension decoding issue
Fixed an issue that could prevent the LDAP SDK from properly
decoding an X.509 certificate extension that made use of a general
names element with a directoryName component.  This did not affect
the LDAP SDK's ability to use certificates containing such
extensions for TLS communication, but it did prevent the
manage-certificates tool from displaying detailed information about
the extension when listing a certificate.