Refresh JAAS configuration before GSSAPI login
Fixed an issue that could interfere with GSSAPI authentication in
cases where a JAAS login module configuration was loaded and cached
by the JVM before all the appropriate system properties had been set
for the login attempt.  The configuration is now refreshed before
attempting to create the login context so that the system property
changes will be applied.