Add a collect-support-data command-line tool
Added a new collect-support-data command-line tool that can be used
to obtain a range of useful information about an instance of the
Ping Identity Directory Server (or related product, like the
Directory Proxy Server, Synchronization Server, or Data Governance
Server) that can help support personnel troubleshoot problems or
better understand the server configuration and the environment in
which it is running.  All of the data collection logic is actually
provided by server-side code, but the tool is included in the LDAP
SDK so that it can be invoked from a client system that does not
have a full server installation.