collect-support-data extended operation updates
Updated the collect support data extended request to make it
possible for the client to indicate the name that it intends to use
for the support data archive file.  Updated the archive fragment
intermediate response to include the name the server used for the
support data archive file.  Updated the collect-support-data
command-line tool so that it no longer requires the outputPath
argument to specified when running with the useRemoteServer option.