Update collect-support-data time range handling
Updated the collect-support-data tool to make a minor change to the
code used to parse values of the logTimeRange argument.  The
server-side version of the code has a not-so-documented feature in
which you can provide the names of rotated log files instead of
well-formed timestamps.  That can't be handled in the LDAP SDK,
especially when operating against a remote server, so if the
useRemoteServer argument is provided, then any time range values
provided must be strictly compliant with one of the supported time
stamp formats.  But when not using a remote server, the tool will
now pass the argument value through to the server-side version of
the code to let it perform the validation.